refactor: 리플레이 컨트롤러 분리
